Fix lint findings: extract route page components, stable row ids
