+++ web/src/main.tsx
@@ -1,20 +1,24 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import { createRouter, RouterProvider } from '@tanstack/react-router'
+import { StrictMode } from "react"
+import { createRouter, RouterProvider } from "@tanstack/react-router"
+import { createRoot } from "react-dom/client"
 
-import './index.css'
-import { routeTree } from './routeTree.gen'
+import "./index.css"
+import { routeTree } from "./routeTree.gen"
 
 const router = createRouter({ routeTree })
 
-declare module '@tanstack/react-router' {
+declare module "@tanstack/react-router" {
   interface Register {
     router: typeof router
   }
 }
 
-createRoot(document.getElementById('root')!).render(
+const root = document.getElementById("root")
+if (!root) {
+  throw new Error("missing #root element")
+}
+createRoot(root).render(
   <StrictMode>
     <RouterProvider router={router} />
-  </StrictMode>,
+  </StrictMode>
 )
