+++ web/vite.config.ts
@@ -1,20 +1,20 @@
-import { fileURLToPath } from 'node:url'
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import tailwindcss from '@tailwindcss/vite'
-import { tanstackRouter } from '@tanstack/router-plugin/vite'
+import { fileURLToPath } from "node:url"
+import tailwindcss from "@tailwindcss/vite"
+import { tanstackRouter } from "@tanstack/router-plugin/vite"
+import react from "@vitejs/plugin-react"
+import { defineConfig } from "vite"
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [
     // tanstackRouter must come before the react plugin
-    tanstackRouter({ target: 'react', autoCodeSplitting: true }),
+    tanstackRouter({ target: "react", autoCodeSplitting: true }),
     react(),
     tailwindcss(),
   ],
   resolve: {
     alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url)),
+      "@": fileURLToPath(new URL("./src", import.meta.url)),
     },
   },
   server: {
@@ -22,13 +22,13 @@ export default defineConfig({
       // The SPA calls /sessions/:id same-origin; in dev, forward it to the Rust
       // Lambda (cargo lambda watch, or API_PROXY_TARGET). In production the
       // Lambda serves the SPA, so the same path is already same-origin.
-      '/sessions': {
-        target: process.env.API_PROXY_TARGET ?? 'http://localhost:9000',
+      "/sessions": {
+        target: process.env.API_PROXY_TARGET ?? "http://localhost:9000",
         changeOrigin: true,
       },
       // '/invoice' also matches the admin '/invoices' paths (prefix match)
-      '/invoice': {
-        target: process.env.API_PROXY_TARGET ?? 'http://localhost:9000',
+      "/invoice": {
+        target: process.env.API_PROXY_TARGET ?? "http://localhost:9000",
         changeOrigin: true,
       },
     },
